Fix performance decrease
Added appearance property that toggles between using the
dayCircleColorSelected property and the current circle color for the
selected days. Even if the default custom day circle color was being
clearColor, the blocks that were setting it were still being called by
default, so I added another appearance property that defaults to NO and
needs to be set to YES not to ignore the custom color blocks.