Add API for users to check and be notified of scrolling state
It currently involves a lot of work to properly update the calendar
asyncronously (as encouraged in the README for long-running update
tasks) because you have to avoid reloading the calendar manager while
either scroll view is still scrolling. This change adds `scrolling`
properties to the menu and horizontal content views so that users can
check the current state, as well as adds two more delegate methods
called when the content scroll view finishes scrolling. This also occurs
when the content scrolling is triggered by menu view scrolling. This
means that users can update their data cache when
calendarDidLoad*Page is called and then reload the calendar when the new
calendarDidFinishScrollingTo*Page is called.