Add zeroPaddedDayFormat property to JTCalendarSettings
Format of the day of the month displayed by JTCalendarDayView can now
be set without leading zero. Single digit days can be displayed as "1"
instead of "01". This is usually the preferred format in the USA.

The date formatter's format is not directly customizable because only
the 2 formats "dd" and "d" make sense in this context.

The default behaviour (zero-padded format) has not changed. However
one drawback of the current implementation is that it's not possible
to toggle between the 2 formats once the calendar has been rendered.
This is because, for performance reasons, the date formatter is cached
in a static variable instead of being recreated every time. It's
unlikely that someone would want to switch between the 2 formats on the
fly so it should be good enough for now.

Fixes gh-283